Move TracingUtils to support package.
